backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
motor==3.3.2
pymongo==4.6.2
openai==1.14.0
python-dotenv==1.0.1
pydantic==2.6.4
python-multipart==0.0.9
aiohttp==3.9.3
arxiv==2.1.0
requests==2.31.0
langchain==0.1.16
langchain-community==0.0.33
langchain-openai==0.1.3
langchain-core==0.1.43
langsmith==0.1.47
